Split out Mac version
