+++ BnbnavNetClient.Mac/AppDelegate.cs
@@ -1,0 +1,15 @@
+namespace BnbnavNetClient.Mac;
+
+[Register("AppDelegate")]
+public class AppDelegate : NSApplicationDelegate
+{
+    public override void DidFinishLaunching(NSNotification notification)
+    {
+        // Insert code here to initialize your application
+    }
+
+    public override void WillTerminate(NSNotification notification)
+    {
+        // Insert code here to tear down your application
+    }
+}
