+++ BnbnavNetClient.Mac/Assets.xcassets/Contents.json
@@ -1,0 +1,6 @@
+﻿{
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  }
+}
