+++ BnbnavNetClient.Mac/BnbnavNetClient.Mac.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk">
+    <PropertyGroup>
+        <TargetFramework>net7.0-macos</TargetFramework>
+        <OutputType>Exe</OutputType>
+        <Nullable>enable</Nullable>
+        <ImplicitUsings>true</ImplicitUsings>
+        <SupportedOSPlatformVersion>10.14</SupportedOSPlatformVersion>
+    </PropertyGroup>
+    <ItemGroup>
+      <ProjectReference Include="..\BnbnavNetClient\BnbnavNetClient.csproj" />
+    </ItemGroup>
+    <ItemGroup>
+      <PackageReference Include="Avalonia.Desktop" Version="11.0.0-preview4" />
+    </ItemGroup>
+</Project>
