+++ BnbnavNetClient.Mac/Main.cs
@@ -1,0 +1,17 @@
+using Avalonia;
+using Avalonia.ReactiveUI;
+using BnbnavNetClient;
+using BnbnavNetClient.Desktop.TextToSpeech;
+using BnbnavNetClient.I18Next;
+using BnbnavNetClient.Settings;
+
+AppBuilder BuildAvaloniaApp()
+    => AppBuilder.Configure<App>()
+        .UsePlatformDetect()
+        .LogToTrace()
+        .UseReactiveUI()
+        .UseI18NextLocalization()
+        .With(new MacTextToSpeechProvider())
+        .UseSettings(new SettingsManagerJsonFile());
+        
+BuildAvaloniaApp().StartWithClassicDesktopLifetime(args);
