+++ BnbnavNetClient.Mac/TextToSpeech/MacTextToSpeechProvider.cs
