+++ BnbnavNetClient.sln
@@ -18,6 +18,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BnbnavNetClient.I18Next", "BnbnavNetClient.I18Next\BnbnavNetClient.I18Next.csproj", "{F9360BD6-A251-478D-BDDA-6B57611055F0}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BnbnavNetClient.Mac", "BnbnavNetClient.Mac\BnbnavNetClient.Mac.csproj", "{10F2F7A6-C407-4D8B-9691-2B4624324AFD}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -40,6 +42,10 @@ Global
 		{F9360BD6-A251-478D-BDDA-6B57611055F0}.Release|Any CPU.Build.0 = Release|Any CPU
 		{F9360BD6-A251-478D-BDDA-6B57611055F0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{F9360BD6-A251-478D-BDDA-6B57611055F0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{10F2F7A6-C407-4D8B-9691-2B4624324AFD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{10F2F7A6-C407-4D8B-9691-2B4624324AFD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{10F2F7A6-C407-4D8B-9691-2B4624324AFD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{10F2F7A6-C407-4D8B-9691-2B4624324AFD}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
